+++ source/tools/DumpBin/DelayImport.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void PrintOneDelayImport(PBYTE Data, DWORD Size, PIMAGE_DELAYLOAD_DESCRIPTOR DelayImportDirectory);
-DWORD DelayImport(PBYTE Data, DWORD Size);
+void PrintOneDelayImport(PBYTE Data, LONGLONG Size, PIMAGE_DELAYLOAD_DESCRIPTOR DelayImportDirectory);
+DWORD DelayImport(PBYTE Data, LONGLONG Size);
